+++ packages/ui/src/components/TeamCard/teamMembers.ts
@@ -1,11 +1,13 @@
-import { TeamMember } from '../TeamCard';
-import avatar from './avatar.png';
+import { TeamMember } from './TeamCard';
+
+const imageUrl =
+  'https://images.unsplash.com/photo-1595152772835-219674b2a8a6?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=800&q=80';
 
 export const members: TeamMember[] = [
   {
     id: 1,
     name: 'Ryan Buckley',
-    image: avatar,
+    image: imageUrl,
     profileUrl: '/',
   },
   {
@@ -16,7 +18,7 @@ export const members: TeamMember[] = [
   {
     id: 3,
     name: 'Ryan Buckley',
-    image: avatar,
+    image: imageUrl,
     profileUrl: '/',
   },
   {
@@ -27,7 +29,7 @@ export const members: TeamMember[] = [
   {
     id: 5,
     name: 'Ryan Buckley',
-    image: avatar,
+    image: imageUrl,
     profileUrl: '/',
   },
   {
@@ -41,7 +43,7 @@ export const mentors: TeamMember[] = [
   {
     id: 7,
     name: 'Ryan Buckley',
-    image: avatar,
+    image: imageUrl,
     profileUrl: '/',
   },
 ];
